requirements.txt
pywinusb==0.4.2
requests==2.25.0
python-dotenv==0.15.0